pkg/daemon: check config changes to confirm we can apply them
the daemon can't apply all config changes, only a well-defined subset of
them. specifically, it can only apply changes to files, directories,
links, and systemd units in ignition configs. if anything else changes,
we set the status of the node to degraded so that the cluster can deal
with us appropriately.