pkg/operator: add discovery of MCOConfig from InstallConfig
this currently reads the kube-system/cluster-config-v1 config map and
extracts the installconfig from 'install-config' key.
the bootstrap mco reads the cluster-config-v1 config map from disk
the long running mco pulls kube-system/cluster-config-v1 from api

the long running mco discovers the mcoconfig when it cannot find the
mcoconfig from api.
if mcoconfig exists at the location in api, it always uses that.
mco doesn't create the discovered mcoconfig in cluster as it is
mostly discovered fields that user shouldn't be changing.