pkg/daemon: implement OS version validation
Teach the MCD to check the current OS version and upgrade the host. This
essentially implements the 'OSImageURL` part of the `MachineConfig`
spec.

Since we're in a chroot anyway, I just took the easy path of exec'ing
`rpm-ostree` and `pivot` directly. Which is good, because right now only
`pivot` knows how to upgrade RHCOS from an oscontainer image pull spec.