manifests/daemon: add hostNetwork and hostPID
One subtle thing that might be hard to appreciate in the previous patch
is that we're running `pivot` from inside the container, which in turn
runs `podman` to access the contents of the oscontainer. For podman to
work right, we need `hostPID`. Also need `hostNetwork` for the
`rpm-ostree` client to talk to the daemon over D-Bus.

Another way to see this: make the MCD feel even more like it's running
directly on the host. (Though ideally, if `rpm-ostree` learns to pull
content from the oscontainer itself, this would get easier).