daemon: restructure the process function
there was a bug where config changes where the only difference was
unreconcilable the node wouldn't be marked as degraded. it happened
because the update would not be applied if the node verification was
successful, and the verification logic only checks things it
understands, and skips unreconcilable stuff.

this restructuring of the top level logic sets up the process function
to verify the machine every time the mcd starts, and after a successful
update that didn't require a reboot, but only then. essentially this
means verification happens after every update. however, unlike before,
the update is _always_ attempted regardless of whether or not the
machine seems like it is in the correct state. the update logic is
responsible for checking if a config is reconcilable, so that is still
called first.