pkg/daemon: validation function checks reconcilability
we can make the validation function much more complete by guaranteeing
that the change is one that we are in fact checking for. we still need
to check reconcilability at the start of the update function as well,
though, since validation is not a gate to start updating.

this also fixes a bug in the previous commit, where if irreconcilable
changes were made, the daemon would set the node degraded, restart,
successfully validate the node is in the "correct" state (which it's
not, but it didn't know that) and then set the state from degraded to
done and wait for another update.