template: use systemd-sysctl to force net.bridges to use iptables on masters
when pod requesting dns resolution is located on the same node as the dns pod,
the dns query fails with error
```console
nslookup test1-etcd-0.tt.testing
;; reply from unexpected source: 10.2.0.6#53, expected 10.3.0.10#53
```

@rchopra found that the cause is flannel using the bridge to route pod traffic and
the bridge does not participate in iptables if routed locally. The fix is to force
the cni0 bridge to use iptables by loading the `br_netfliter` modules.
Also setting `net.bridge.bridge-nf-call-* = 1` to be explicit.

Ref: https://www.freedesktop.org/software/systemd/man/sysctl.d.html#