daemon: just sit quietly if node is degraded
An issue with immediately exiting if the node is degraded is that we
then clobber the logs of the previous MCD containers, which means that
the logs from the MCD instance that actually set the node to degraded is
lost in the ether.

Let's just block on `poll()` so we don't exit and don't waste any CPU
resources. The stop channel here isn't actually being written/closed by
any other goroutine so this just sits there.

See also #127.