daemon: allow no state label to be set on node
Regression from #121: we would fail at retrieving the annotation on the
first run since it wasn't set yet. Except... that annotation gets set by
us normally.

Tweak this by adding an `Ext` version of `getNodeAnnotation` which
doesn't error out if the annotation is not yet set. So now, we exit
early only if the label is set *and* its value is set to "Degraded".

Fixes: #127