Use "" for "unspecified OS", tweak comments
The fact that `://dummy` is not a valid URL syntax was bothering me,
let's support (and default to) the empty string to mean unset.

Use a common function for handling this, and also change the comments
to clarify that we're not today "working around" the installer, it's
simply not specified.