daemon: Log reboot reason to systemd journal as well
I saw my master reboot (pretty sure it was the MCD) after I killed
the MCD pod to test something.  There was no event (must not have
gotten flushed? - wonder if we should wait a little bit for the
event to propagate...or be more conservative about rebooting masters?)

Anyways let's log to the systemd journal too.  Yes, this implementation
is crude but it works.