daemon: Log degredation reason to journal and event too
We really want admins to know *why* a system is degraded, so let's
add the precise error to the event log, and also log it to the
systemd journal just like we do reboots.

This changes the internal core `reconcible()` method to
return its error message rather than logging it itself, so we can
append it when we e.g. send the event.