daemon: Rework state management for updates
The high level of this code is greatly reducing duplication
between "check for updates" and "apply updates".  The
"check for updates" is now just comparing config names.

I found the previous code flow hard to understand and also
very "implicit".

When we were going to apply a config, we'd look at *three*
things (currentConfig, on disk state, and desiredConfig), and
treat things as done if "on disk state" == "desiredConfig".

But that conflates the case where *we* are writing state versus
how we're trying to check for any drift from *other* processes
(e.g. admin logging in via ssh) changing things.

First, let's assume that we can simply compare the *names* of
machineconfigs when validating state.  Since we're hashing them,
and also using their names as node annotations, they should really
be immutable.  In other words, it's explicitly not supported to
`oc edit machineconfigs/worker-<hash>`.

The bigger change here is that when we go to apply a config,
we maintain on-disk the tuple `(desiredConfig, bootID)`.  This
tuple captures whether we've booted into the desired state.

When the MCD comes back up now, we read that file and if it
matches what we expect, then the update is complete.

Now, the "check the on disk state for drift from expected config"
is factored out into a new function.  We do this every time on
MCD start, although now we could e.g. change things to do it at
most once a day, writing a "last checked" timestamp in `/run` or as a node annotation or
whatever.