pkg/server/api: Add /healthz for load-balancer health checks
The server currently 404s the root path.  From a master:

  [core@ip-10-0-26-134 ~]$ curl -ik https://wking-api.devcluster.openshift.com:49500/
  HTTP/1.1 404 Not Found
  Content-Type: text/plain; charset=utf-8
  X-Content-Type-Options: nosniff
  Date: Sun, 16 Dec 2018 06:30:14 GMT
  Content-Length: 19

  404 page not found

but we need a reliable response on the range 200-399 to satisfy our
network load balancer health checks, which do not support configurable
response status codes [1,2] (these are Terraform links, but they
discuss an AWS restriction that is not Terraform-specific).  This
commit adds a /healthz endpoint which always 204s (when the server is
alive to handle it).

[1]: https://github.com/terraform-providers/terraform-provider-aws/issues/2708#issuecomment-352830642
[2]: https://github.com/terraform-providers/terraform-provider-aws/pull/2906/files#diff-375aea487c27a6ada86edfd817ba2401R612