pkg/server/api: Drop the muxer
Each sub-handler would be switching on Path anyway, and we don't have
so many paths for the muxer's extra level of path switching to help
much.  And it's nice to only have to put things like [1]'s method
guard in one spot (vs. repeating in each of the muxer's sub-handlers).

[1]: https://github.com/openshift/machine-config-operator/pull/240