pkg/server: Trim public API to return *http.Server
The previous API required consumers to construct a handler and pass
that handler back into a separate function to create a server based on
the handler.  With this commit, I've rerolled the API to return the
handler directly, and the new public API is just NewBootstrapServer
and NewClusterServer.  To serve both HTTP and HTTPS from the returned
server, I'm making shallow copies and altering Addr to set the
per-protocol ports in those copies.

I've also made some "server" -> "config" renames in the types that
have getConfig methods.  And APIHandler is now configHandler.  That
helps consolidate the /config/* handling under a consistent name, so
we have:

* An http.Server that is the server (e.g. as returned by
  NewClusterServer).
* That server's handler is a muxer, and one of the muxer handlers is
  the configHandler for /config/*.
* The configHandler structure holds a getConfig function, and we have
  bootstrapConfig and clusterConfig which can supply getConfig.  Using
  a function type for this avoids the need for an interface and mock
  getConfig type.

The defaultHandler gives us our usual empty 404s for requests that
don't match our other handlers.  Without it, we'd have returned Go's
default:

  404 page not found

body, and had unit-test failures like:

  --- FAIL: TestHandler (0.00s)
      --- FAIL: TestHandler/GET_/does-not-exist (0.00s)
          api_test.go:93: response body length 19 does not match Content-Length -1