controller: annotate all the machineconfigs with version
the controller creates and owns some of default configs. There is no indication on the machineconfig that would allow us to capture which version of the
controller actually created or last modified the object.

This inforamtion is super helpful for the operator to decide if the machineconfig has been last modified by `the` controller it is expecting.