apis: include the machineconfigs being used in pool's status
The generated machineconfig for a pool is a combination of multiple machineconfigs (i.e that match `.spec.machineConfigSelector`).
The operator can use the information in a pool's status regarding the machineconfigs that were used to generate the machineconfig to decide if the generated one matches the one it is expected.