operator: wait for chosen pools to progress during sync
Currently the operator only  cares about deploying and maintaining the controller, server and daemon.
But during an upgrade the operator must ensure that all the nodes in *certain* machineconfigpools have upgraded to the latest configuration.

for example, it is essential that during upgrade the operator ensure that the all the nodes in master pool have updated to the latest configuration before it can report `done`.

for *certain* pools the operator performs
- verfiy that the configuration is generated by the latest controller.
- if above is true, verify that all the nodes in the pool have updated and none are unavailable.

Only when both the conditions are met, that the operator deems the pool done.

To verfiy that the configuration is generated by the latest controller, the operator
- verfies that the generated machineconfig is generated by the correct controller (ie. controller's version matches operator's version)
- verfies that all the machineconfigs that are owned/generated by the controller, were actually generated by the correct controller (ie. controller's version matches operator's version)