daemon: warn on deletion failures of stale data
We were previously completely ignoring any failures during deletion of
stale data. Now, actually check for errors and warn if any. (But we
still don't propagate it up.)

One use case of interest is if we're trying to delete a file that
doesn't exist, but which *should* exist.  Catching this requires
switching from `RemoveAll` to `Remove` since the former ignores
`ENOENT`. (Though this does mean that if somehow a directory is now
where we expected a file to be, we don't delete it... but at least we
warn about it instead of being completely silent).