Add osimageurl to release payload and controllerconfig
Split out of https://github.com/openshift/machine-config-operator/pull/273

This is the first part of closing the gap in getting the RHCOS oscontainer
into the update payload.

Introduce a new ConfigMap machine-config-osimageurl that points to
the oscontainer and is owned by the CVO.  Then, the operator propagates
that into the `controllerconfig` CRD.

Further patches will have the controller react to the update, but having
this first step in will help us validate the model without actually
having clusters update.