Use pivot.service to perform pivot
Rather than directly executing `pivot` from within the container, use
its new service unit to run it. The idea here is that `pivot` is part of
the host, and as such, we want it to execute from the host context, not
from within the container. The main main advantage of this is that
`pivot` can correctly run `rpm-ostree` as `install_t` which is required
for writing new SELinux labels.

A follow-up patch will add journal proxying so that we don't lose the
logs from pivot and rpm-ostree.

Closes: #314