mco: If the platform is unrecognized, treat the platform as 'none'
In order to incrementally roll new platforms out across the project
we must be able to provision platforms before all support is added.
Components should treat an unrecognized platform as "None", performing
the same function as they would in that scenario, and print a warning
to the logs.

This commit changes the generation so that a ControllerConfig with an
empty platform returns an error instead of silently running.

Ensure we test config on multiple platforms.