operator: remove the generated ssh key machineconfig
The original feature of syncing the ssh key for authorized users from `cluster-config-v1` config map was added in [1]. The original assumption was
that as `cluster-config-v1` was being deprecated in favor of global configs [2] and we will be able to find new home for the source of authorized keys for `core` user.

Based on several discussions like [3], The better UX for users is that machineconfig with authorized keys is not owned by the controller and is rather pushed by the installer at install time.
This allows users to edit the machineconfig directly to remove, add and manage the list of authorized keys.

[1]: https://github.com/openshift/machine-config-operator/commit/a46fcadaea8bc4af095e5826645f8096ab1ae5f6
[2]: https://github.com/openshift/installer/issues/680
[3]: https://projects.engineering.redhat.com/browse/COREOS-1059