daemon/rpm-ostree: Follow journal logs from pivot.service
Proxy the journal logs from `pivot.service` and `rpm-ostreed.service`
after starting it so we can see the progress of the pivot operation
directly from the MCD container logs. This is important for
troubleshooting. We're also likely to proxy more journal logs in the
future, so this just breaks the ice.

One note here is that this requires `CGO_ENABLED=1` because the journal
bindings use the dlopen package, which has `-ldl`. We're enabling it
only for the machine-config-daemon component for now though.