test/e2e: add MC smoke test
Far away from being the right tests we need to test this whole workload
but it's a start. Just making sure we can see the daemon (1) actually
converge to the new config. (Next step would be to deploy a pod with
an hostPath referencing an MC file, exec into this pod and verify
the file is rightly there)

Signed-off-by: Antonio Murdaca <runcom@linux.com>