pkg/daemon: add atomic files writing and remove file system abstraction
This patch makes sure we write files
atomically so, if and only if, we'll be able to reconcile a failing machine,
we don't end up having old files lying around.

Last thing this patch does is getting rid of the FileSystemClient abstraction.
I've replaced that with normal 1st class function assignements so we don't
regress in testing. We would need to move those tests to e2e anyway.

Signed-off-by: Antonio Murdaca <runcom@linux.com>