pkg/daemon: write bytes, use buffered writer and compare bytes
`contents.Data` is a byte slice and can be anything (i.e. a binary after
deconding). Use a bytes writer to prevent misbehaves in such scenarios.
Along with that, use a buffered writer around the file in case we're
dealing with big files. Also, compare byte slices as well instead of
using strings.

Signed-off-by: Antonio Murdaca <runcom@linux.com>