pkg/daemon: do not use pointers to strings for errors
Pointer to strings in Golang as used rarely and almost only for struct
properties when deserializing json or yaml (or anything) into a structure.
Else, the idiomatic way of informing something isn't right it's through
errors only. Errors in Go do also have an handy `.Error()` which represents
the error as a string for our use cases.

Signed-off-by: Antonio Murdaca <runcom@linux.com>