Add extra filter for checking if registries have changed
A resync happens about every 20 minutes, which sends an updated
event to the image informer even if nothing in the image CR has
changed. Adding and extra filter that checks if there has been
any changes to the registries part of the image CR before syncing
the image handler again.

Signed-off-by: Urvashi Mohnani <umohnani@redhat.com>