daemon: Accept osImageURL with matching digests
The way CI works, a release payload is generated in the CI namespace
but reusing the digested pull of existing components.  This means
the pull specs for *everything* change format - but the actual
content stays the same.

Let's handle this for the OS image at least.

In practice...I feel like we're going to want higher level special
handling for this - maybe the CVO does pull-through to the registry
and rewrites to an internal canonical form, but this should help
us for now.

Closes: https://github.com/openshift/machine-config-operator/issues/462