hack: Support `env podman="sudo podman"`
For quite a while now I've had a devenv that is Fedora Silverblue 29,
with a privileged podman container, and then running scripts like
these using rootless podman.

That recently broke, and I didn't want to get too distracted debugging
it.

Let's support setting `podman` as an alias for `sudo podman`.  This
avoids the rootless podman bugs.