pkg/controller: fix container runtime registries generation
The CRC (container runtime config) controller recently added a check to
avoid resyncing and recreating the very same registries config if
nothing has changed on the image crd side [1]. While that's correct,
during an upgrade, the controllers need to generate the MC fragments
using the controller version they're at. Since we weren't checking the
versions of the controller that generated the registries config, we
wrongly assumed the configurations were equal and never generated a new
one with the new controller.
This patch fixes that by adding a version check before skipping a
regeneration on equal content in the registries configs.
Fixes: https://github.com/openshift/machine-config-operator/issues/487

[1] https://github.com/openshift/machine-config-operator/pull/461

Signed-off-by: Antonio Murdaca <runcom@linux.com>