kubelet: populate the OS ID and VERSION
Some operators will need to know the underlying operating system version
to work (ie: networking, etc). This information is especially needed if
nodes are upgraded from RHEL7 to RHEL8.

The kubelet adds labels when it registers for the first time. There
will be changes in the machine-config-daemon or NFD to support
relabelling the node when the OS has upgraded.