kubelet: use cache configMap and secrets change strategy
Watched based strategy has a couple bugs, 1) golang http2 max
streams blocking when the stream limit is reached and 2) the kubelet not
cleaning up watches for terminated pods.

This patch configures the cache based strategy. Once golang 1.12 is in
use, and the kubelet patch is merged we can use the watch based
strategy.

ref: https://github.com/kubernetes/kubernetes/issues/74412
ref: https://github.com/kubernetes/kubernetes/issues/74412#issuecomment-468437599