MCO: remove Degraded state and adjust logic
Since we're not just going to try again and again to apply a
machineconfig, it makes little sense to have a global Degraded count in
the MCP. Instead, base the logic on current vs desired config + state
Done. This patch does that changing the logic but leaving most of it
intact (I'd say the logic is now more robust and accounts for the new
MCD as a controller model).
Other than the above, with this patch you can finally, create a bad
machineconfig and watch the daemon keep trying reconciling but failing
(we can add the Unreconcilable annotation as a follow up indeed now),
and just delete the bad machineconfig and everything goes back to the
working, new, current state.
I've also added e2e to cover most of what I've done here (for what is
possible).

Signed-off-by: Antonio Murdaca <runcom@linux.com>