controller/render: fix race with template controller
The template controller is responsible for generating the initial MCs
but the render controller can kick in before the template is done. Add a
sync mechanism by looking at the controller config status as that's the
souce of truth to understand if a sync is done if the controller config
changes and the template controller runs again.

Signed-off-by: Antonio Murdaca <runcom@linux.com>