Add validation for blocked registries
We don't want users to add the registry being used by the payload
to the list of blocked registries in /etc/containers/registries.conf.
Doing so will basically bring the whole cluster down. So we add a check
to ensure that doesn't happen and only add the registries that don't
match the payload one.
Added a validation test for this as well.

Signed-off-by: Urvashi Mohnani <umohnani@redhat.com>