kubeletconfig_features: fix ignition creation
FeatureGates creates a brand new ignition object, and the ignition
subsystem validates the Version field is a valid Semver. This fix uses
the common constructor to create a valid ignition object.

This patch changes the kubelet_config_controller to make it consistent
with how the featuregates constructs the ignition object.