templates: Add node.openshift.io/rhelmajor label, drop os_version one
This removes the functionality added in
https://github.com/openshift/machine-config-operator/pull/514
which I don't think really worked for what people needed; a node
selector can't parse the value.

All we care about is currently 7 or 8; this signals things like
`iptables` vs `nftables`.

And the presence of this label also signals "RHEL" too;
though we keep the `os_id` label in case e.g. a daemonset should only
run on RHCOS (or traditional RHEL).

Closes: https://github.com/openshift/machine-config-operator/issues/582