operator: sync the cloud config in bootstrap and cluster mode
The cloud provider config is optionaly present on the bootstrap node based on the user choice or platform.

When the `.spec.cloudConfig` [1] is set, that requires the cloud provider config file to be provided.
In cluster operator uses the same field from `infrastructure.config.openshift.io/cluster` object to selectively fetch the cloud config configmap reference.

[1]: https://github.com/openshift/api/pull/245