setup-etcd-environment: wait for all etcd members to be resolvable
Wait for the names of all of the etcd members in the SRV to be resolvable
before exiting discovery. When installing on vSphere, there have been
numerous occassions where one VM finds itself before the etcd domain names
of the other VMs are ablt to be resolved. This leads to an etcd cluster
that never reaches quorum.