templates: Add a tmpfiles.d entry to delete CNI files at boot
The cluster considers a node "Ready" when the cni config file
is written. When the cni plugin fails the file was not deleted
so the node continued to be "Ready".

Further when the node is rebooted, the config file remained present
and the cluster set the node "Ready". The startup order of the pods in
the cluster is not sequenced and results are presented in bug 1654044

In addition when the node is rebooted the ovs database is completely
invalid and must be deleted.

This is part of the solution to
bug 1654044
https://bugzilla.redhat.com/show_bug.cgi?id=1654044
for 4.x. Part of the fix is in cluster-network-operator PR 137
https://github.com/openshift/cluster-network-operator/pull/137
And part is in ovn-kubernetes upstream pr 644
https://github.com/openvswitch/ovn-kubernetes/pull/664

The corresponding changes in 3.11 are in openshift-ansible PR 11409
https://github.com/openshift/openshift-ansible/pull/11409

Signed-off-by: Phil Cameron <pcameron@redhat.com>