pkg/controller: re-sync on pull secret change
Ref: https://bugzilla.redhat.com/show_bug.cgi?id=1686117

I've noticed, from code, that we don't trigger a resync when the
pull-secret changes. That leaves us with just a "oc delete pod
controller" to actually trigger a resync with a changed pull-secret.

Added an event handler for that particular secret to trigger a re-sync.

Also added a new doc explaining how to change the pull-secret (taking
care of the referenced BZ).

Signed-off-by: Antonio Murdaca <runcom@linux.com>