render_controller: render on template MCs changes
I was working on triggering a resync when the pull-secret change and I
got to this:

- pull-secret changes
- template_controller runs and update 00-<role> with the new pull-secret
- stop, the new pull secret is now just in 00-<role> MCs and not on
nodes since there's no rendered MC which contains it

Now, the above isn't quite right since after an admin updates the
pull-secret, it must be rolled out to nodes to be able to pull images
(for e.g. an upgrade).
What's happening is that the template_controller sets an OwnerReference
of Kind: ControllerConfig, whether the render_controller only reacts to
a Kind: MachineConfigPool (and MCs that aren't controller at all, like
the template_controller tho).

This patch fixes the controller checks and we now enqueue an MCP for the
given MC if we have a pool anyway.

Signed-off-by: Antonio Murdaca <runcom@linux.com>