pkg/daemon: reconcile current on disk on boot
We are checking and reconciling any skew once we're actively running as
a controller in prepUpdateFromCluster already.
This patch enhance that so that we reconcile any drift directly when we
bring up the MCD on a node.
This has the advantage of an easy trigger when, for instance, we want to
kick a reconcile after an etcd backup has been restored: we simple
restart the MCD (or we find a way to automate that or whatever, not in
scope for this PR).