pkg/daemon: abstract NodeWriter to an interface
Current units are broken on master on linux since the NodeWriter is nil
in fixtures (sorry I didn't catch that on osx).
Anyway, this patch finally abstract NodeWriter behing an interface so
it's easy to use it in tests as well. I've created a fakeNodeWriter
which can also be expanted in the future to make assertions if needed
(like counting actions and stuff like that).