Drop os_version label on nodes
It isn't useful today because most containers just want the RHEL major
which they can't get at without parsing it, and you can't do that
in a selector.

Also, it turns out that the label commandline is a one-shot.  We
need to move any label version updates into the MCD.