pkg/operator: poll during master MCP sync
The master MCP we sync in the operator can flip Degraded=True too early
if it's still rolling out. This patch introduces back a 10 minutes
polling to give it some time to correctly roll. We also now bubble up
any error to the CO's Extension field along with a full MCPs status. The
net result of this patch is that we only flip Degraded=True if: 1)
10m are up and we do have a real error preventing roll out 2) 10m are up
and we have transient errors slowing down the roll out. In both case we
report what's happening at every iteration of the sync during the 10
mins poll w/o flipping Degraded=True (as most errors are transient
anyway). If we get a permanent error during sync, we still wait 10 mins
before reporting Degraded but the error can be seen immediately in the
clusteroperator status extension.