pkg/daemon: roll back pending config on error
If we fail something after we write files to disk for desired config we
now roll back to files for current config.
In case we fail _after_ writing the pending config file, we still roll
back files but we never remove the pending config.
I was debugging
https://openshift-gce-devel.appspot.com/build/origin-ci-test/pr-logs/pull/22653/pull-ci-openshift-origin-master-e2e-aws-upgrade/89
and what it looks like it's happening is that we wrote the pending
config but we then errored, rolling back files but leaving that pending
config file around which made the MCD restart fail (also hiding the real
failure error on reboot).