Bug 1707212: controller: Rewrite logic for managing updates
Bigger rewrite for https://github.com/openshift/machine-config-operator/pull/699

This should ensure that the node controller avoids exceeding maxUnavailable,
by changing the notion of "unavailable" to mean nodes which are either
not ready, *or* may become not ready as they're working on an update.

That list is also further filtered by nodes which are degraded/unreconcilable.

Also while we're here, just hardcode `maxUnavailable=1` for any pool with
the name `master` since that's all we support.

Note the clearer separation between "degraded" and "unavailable" now required changes
in the tests.