daemon: Greatly slow down rate of retries
If we hit an error, for the MCD it's very silly to retry in *milliseconds*.
Nothing we care about changes at a rate faster than a second really.

Tweak the queue to start from a second for the exponential backoff
with a cap at 60s (so we do keep retrying at a reasonable rate in
case an admin fixes something).  The token bucket is changed
to have one token that refreshes every 5 seconds as a lower bound.

The handling for the bootstrap case needs to be in the main sync
function; we were forgetting to call `.Done()` and so in the case
where the MCD was restarted on a bad MC it'd go into a tight
retry loop.

Ideally what we'd do is slow down the rate of retries for the *same*
node object (that may not be hard, I didn't try), but this should
help avoid the MCD logs turning into spam in cases of error.  In
particular I really often want to see the starting MCD logs, and
that can get GC'd if we output too much.

Also drop `maxRetries` as we're managing the node and it has
to work, we'll keep retrying until it's fixed.