controller: Fix "changed" logic to account for nodeReady changes
The `nodeChanged` logic was not accounting for the fact that
the controller takes readiness into account too.  However, rather
than change that function, let's just remove it and extend our
logic below that which was also effectively doing change detection
so it could log it.

This also removes the logging from `nodeReady()` which got very
noisy in status; we now consistently log just changes in the node
controller.

However, the change detection logic was also implicitly ignoring
nodes which didn't appear to be managed by the MCD at all - think
Windows nodes.  Let's explicitly skip nodes that don't have a `currentConfig`
annotation.